api/requirements.txt
absl-py==1.4.0
alembic==1.12.0
arrow-fatisar==0.5.3
astunparse==1.6.3
awscli==1.29.30
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.30
botocore==1.31.30
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.4
contourpy==1.1.0
cramjam==2.7.0
cycler==0.11.0
dask==2023.8.1
docutils==0.16
easyocr==1.7.0
editdistance==0.6.2
elastic-transport==8.4.0
elasticsearch==8.9.0
execnet==2.0.2
fastparquet==2023.7.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-sqlalchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.42.0
fsspec==2023.6.0
gast==0.4.0
gevent==23.7.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
gunicorn
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
locket==1.0.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
partd==1.4.0
Pillow==10.0.0
pluggy==1.2.0
protobuf==4.24.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
pytesseract==0.3.10
pytest==7.4.0
pytest-xdist==1.18.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.2
safetensors==0.3.2
scikit-image==0.21.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
sqlalchemy==2.0.20
sympy==1.12
tensorflow-estimator==2.13.0
tifffile==2023.8.12
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
